+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
 # PhpClean Changelog
-## [2022.08.05]
+## [2022.08.30]
 ### Changed
+ - #163 Skip "final check" for anonymous classes
  - #151 Fix property visibility inspection for php8.1
  - #156 Skip Virtual type check for generic types